Update to new version of gcloud java Rename packages
