+++ bookshelf/pom.xml
@@ -34,12 +34,12 @@ Copyright 2015 Google Inc. All Rights Reserved.
     <dependency>
       <groupId>com.google.apis</groupId>
       <artifactId>google-api-services-plus</artifactId>
-      <version>v1-rev327-1.21.0</version>
+      <version>v1-rev329-1.21.0</version>
     </dependency>
     <dependency>
       <groupId>com.google.gcloud</groupId>
       <artifactId>gcloud-java</artifactId>
-      <version>0.1.2</version>
+      <version>0.1.3</version>
     </dependency>
     <dependency>
         <groupId>com.google.guava</groupId>
@@ -61,7 +61,7 @@ Copyright 2015 Google Inc. All Rights Reserved.
     <dependency>
       <groupId>joda-time</groupId>
       <artifactId>joda-time</artifactId>
-      <version>2.9.1</version>
+      <version>2.9.2</version>
     </dependency>
     <dependency>
       <groupId>jstl</groupId>
