+++ bookshelf/src/main/appengine/app.yaml
@@ -8,7 +8,6 @@ runtime_config:
 env_variables:
   STORAGETYPE: YOUR-DATABASE-CHOICE
   SQL_DATABASE_URL: jdbc:mysql://HOST-IP:3306/YOUR-DB-NAME?user=USERNAME&password=PASSWORD
-  PROJECT_ID: YOUR-PROJECT-ID
   BUCKET_NAME: YOUR-BUCKET-NAME
   CLIENT_ID: YOUR-CLIENT-ID
   CLIENT_SECRET: YOUR-CLIENT-SECRET
