+++ bookshelf/src/main/java/com/example/managedvms/gettingstartedjava/daos/BookDao.java
@@ -14,24 +14,23 @@
  * limitations under the License.
  */
 
-package com.example.appengine.gettingstartedjava.objects;
+package com.example.managedvms.gettingstartedjava.daos;
 
-import java.util.List;
+import com.example.managedvms.gettingstartedjava.objects.Book;
+import com.example.managedvms.gettingstartedjava.objects.Result;
 
 // [START example]
-public class Result<K> {
+public interface BookDao {
+  public Long createBook(Book book) throws Exception;
 
-  public String cursor;
-  public List<K> result;
+  public Book readBook(Long bookId) throws Exception;
 
-  public Result(List<K> result, String cursor) {
-    this.result = result;
-    this.cursor = cursor;
-  }
+  public void updateBook(Book book) throws Exception;
 
-  public Result(List<K> result) {
-    this.result = result;
-    this.cursor = null;
-  }
+  public void deleteBook(Long bookId) throws Exception;
+
+  public Result<Book> listBooks(String startCursor) throws Exception;
+
+  public Result<Book> listBooksByUser(String userId, String startCursor) throws Exception;
 }
 // [END example]
