+++ bookshelf/src/main/java/com/example/managedvms/gettingstartedjava/daos/CloudSqlDao.java
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package com.example.appengine.gettingstartedjava.daos;
+package com.example.managedvms.gettingstartedjava.daos;
 
 import org.apache.commons.dbcp2.BasicDataSource;
 
-import com.example.appengine.gettingstartedjava.objects.Book;
-import com.example.appengine.gettingstartedjava.objects.Result;
+import com.example.managedvms.gettingstartedjava.objects.Book;
+import com.example.managedvms.gettingstartedjava.objects.Result;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
